@@ -1,6 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/http';
 
 // Components
 import { AppComponent } from './app.component';
@@ -21,7 +21,7 @@ import { VideoYoutubePipe } from './pipes/video-youtube.pipe';
   ],
   imports: [
     BrowserModule,
-    HttpModule
+    HttpClientModule
   ],
   providers: [YoutubeService],
   bootstrap: [AppComponent]
